fix up for changing devices. still working on WMPhoto blob getting downloaded

+++ WoundMapUS/WoundMapUS/Common/Classes/Controller/WMPrintConfigureViewController.m
@@ -647,7 +647,7 @@
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    NSString *cellIdentifier = [NSString stringWithFormat:@"%d-%d", indexPath.section, indexPath.row];
+    NSString *cellIdentifier = [NSString stringWithFormat:@"%ld-%ld", (long)indexPath.section, (long)indexPath.row];
     UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:cellIdentifier];
     if (cell == nil) {
         if ([self indexPathIsPasswordCell:indexPath]) {
